# src/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "FinaMantic",
        "description": "FinaMantic"
    },
    "basePath": "/",
    "schemes": [
        "http",
        "https"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
...
        "/auth/log-in": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Log in system",
                "description": "Log in system",
                "operationId": "log-in",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "$ref": "#/parameters/HeaderVersion"
                    },
                    {
                        "$ref": "#/parameters/HeaderLanguage"
                    },
                    {
                        "$ref": "#/parameters/HeaderPlatform"
                    },
                    {
                        "in": "body",
                        "name": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/login"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Login",
                        "schema": {
                            "$ref": "#/definitions/responsesOk"
                        }
                    },
                    "500": {
                        "description": "Unexpected error.",
                        "schema": {
                            "$ref": "#/definitions/responsesError"
                        }
                    }
                }
            }
        },
...
        "/auth/forgot-password": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Forgot Password",
                "description": "Forgot Password",
                "operationId": "forgot-password",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "$ref": "#/parameters/HeaderVersion"
                    },
                    {
                        "$ref": "#/parameters/HeaderLanguage"
                    },
                    {
                        "$ref": "#/parameters/HeaderPlatform"
                    },
                    {
                        "in": "body",
                        "name": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/forgot_password"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Forgot password in system",
                        "schema": {
                            "$ref": "#/definitions/responsesOk"
                        }
                    },
                    "500": {
                        "description": "Unexpected error.",
                        "schema": {
                            "$ref": "#/definitions/responsesError"
                        }
                    }
                }
            }
        },
        "/auth/reset-pswd": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Reset Password",
                "description": "Reset password",
                "operationId": "reset-password",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "$ref": "#/parameters/HeaderVersion"
                    },
                    {
                        "$ref": "#/parameters/HeaderLanguage"
                    },
                    {
                        "$ref": "#/parameters/HeaderPlatform"
                    },
                    {
                        "in": "body",
                        "name": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/reset_password"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User profile in system",
                        "schema": {
                            "$ref": "#/definitions/responsesOk"
                        }
                    },
                    "500": {
                        "description": "Unexpected error.",
                        "schema": {
                            "$ref": "#/definitions/responsesError"
                        }
                    }
                }
            }
        },
...
        "/users/categories":{
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "All categories list for add goals",
                "description": "All categories list for add goals",
                "operationId": "categories",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "$ref": "#/parameters/HeaderVersion"
                    },
                    {
                        "$ref": "#/parameters/HeaderLanguage"
                    },
                    {
                        "$ref": "#/parameters/HeaderPlatform"
                    },
                    {
                        "$ref": "#/parameters/HeaderAuth"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "all categories list ",
                        "schema": {
                            "$ref": "#/definitions/responsesOk"
                        }
                    },
                    "500": {
                        "description": "Unexpected error.",
                        "schema": {
                            "$ref": "#/definitions/responsesError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "login": {
            "type": "object",
            "required": [
                "email"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "example": "mushrraf@yopmail.com"
                },
                "password": {
                    "type": "string",
                    "example": "12345678"
                },
                "deviceToken": {
                    "type": "string",
                    "example": "simulator"
                },
                "deviceType": {
                    "type": "string",
                    "example": "oppo"
                },
                "deviceId": {
                    "type": "string",
                    "example": "simulator"
                }
            }
        },
...
        "reset_password": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "deviceId": {
                    "type": "string",
                    "description": "Unique identification number of device",
                    "example": "asfd12erf34"
                },
                "email": {
                    "type": "string",
                    "description": "Email",
                    "example": "abc@yopmail.com"
                },
                "password": {
                    "type": "string",
                    "format": "password",
                    "example": "1234"
                }
            }
        },
        "forgot_password": {
            "type": "object",
            "required": [
                "email"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "example": "test@yopmail.com"
                }
            }
        },
...
        "responsesOk": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string"
                },
                "code": {
                    "type": "string"
                },
                "msg": {
                    "type": "string"
                },
                "data": {
                    "type": "object"
                }
            }
        },
        "responsesError": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string"
                },
                "code": {
                    "type": "string"
                },
                "msg": {
                    "type": "string"
                },
                "error": {
                    "type": "object"
                }
            }
        },
...
    },
    "parameters": {
        "HeaderVersion": {
            "name": "X-FinaMantic-Version",
            "in": "header",
            "description": "X-FinaMantic-Version",
            "required": true,
            "type": "string",
            "default": "1.0.0"
        },
        "HeaderLanguage": {
            "name": "Accept-Language",
            "in": "header",
            "description": "Accept-Language",
            "required": true,
            "type": "string",
            "default": "en"
        },
        "HeaderPlatform": {
            "name": "X-FinaMantic-Platform",
            "in": "header",
            "description": "android or ios or web",
            "required": true,
            "type": "string",
            "default": "android"
        },
        "HeaderAuth": {
            "name": "Authorization",
            "in": "header",
            "description": "Authorization",
            "required": true,
            "type": "string"
        },
...
    }
}